fix: コードレビュー指摘事項を対応（セキュリティ・型安全・データ整合性）
- GlobalExceptionHandler: catch-all 500ハンドラーのex.message露出を防止し固定メッセージに変更
- CoffeeBeanDetailRow/Result/Response + CoffeeBeanMapper: tasteIdをレスポンスに追加しテイスト名変更時の評価値リセットを防止
- EditCoffeeBeanModal: テイスト評価のデフォルト値照合をtasteName文字列一致→tasteIdに変更
- CreateCoffeeBeanButtonClient: tastes prop型をTasteListItem[]|null→TasteListItem[]に修正しビルドエラーを解消
- tastes.ts: JSDocをthrow動作と一致するよう修正
- page.tsx/CreateCoffeeBeanButton: fetchTastes().catch(()=>[])でテイストAPI障害時のページクラッシュを防止
- parseTastes.ts: createCoffeeBeanAction/editCoffeeBeanActionのtaste変換ロジックを共通化
- TasteProfileField: 不要な"use client"ディレクティブを削除
