+++ admin-frontend/src/app/(admin)/coffee-beans/[id]/_components/editCoffeeBeanAction.ts
@@ -7,6 +7,7 @@ import {
   type CoffeeBeanFormState,
   coffeeBeanFieldsSchema,
 } from "@/app/(admin)/coffee-beans/_lib/coffeeBeanFormSchema";
+import { parseTastesFromFormData } from "@/app/(admin)/coffee-beans/_lib/parseTastes";
 
 const editCoffeeBeanSchema = coffeeBeanFieldsSchema.extend({
   id: z.string(),
@@ -41,17 +42,7 @@ export async function editCoffeeBeanAction(
 
   const { id, ...fields } = result.data;
 
-  const tasteIds = formData.getAll("tasteIds") as string[];
-  const tastes = tasteIds
-    .map((tasteId) => {
-      const rawValue = formData.get(`tasteValue_${tasteId}`) as string;
-      const evaluationValue = Number.parseInt(rawValue, 10);
-      if (Number.isNaN(evaluationValue)) return null;
-      return { tasteId, evaluationValue };
-    })
-    .filter(
-      (t): t is { tasteId: string; evaluationValue: number } => t !== null,
-    );
+  const tastes = parseTastesFromFormData(formData);
 
   const response = await multipartRequest(
     `/api/admin/coffee-beans/${id}`,
