+++ backend/admin-api/src/main/kotlin/com/mametosho/admin/application/query/result/CoffeeBeanDetailResult.kt
@@ -22,6 +22,7 @@ data class CoffeeBeanDetailResult(
 
     data class TasteResult(
         val id: String,
+        val tasteId: String,
         val tasteName: String,
         val evaluationValue: Int,
     )
